# openapi.yaml
openapi: 3.1.0
info:
  title: AIDecepticon Control Plane API
  version: 0.1.0
  description: API for managing organization-isolated deception deployments, sensors, tokens, incidents, encrypted secrets, audit events, identity domains, and integrations. Operator endpoints use X-AIDecepticon-Organization to select an authorized organization.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8787/api/v1
security:
  - controlPlaneSession: []
  - controlPlaneBearer: []
paths:
...
  /sensor-commands:
    get:
      operationId: listSensorCommands
      summary: List sensor commands, optionally filtered by lifecycle state
      security:
        - controlPlaneSession: []
        - controlPlaneBearer: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { description: Sensor command collection }
  /sensor-commands/{commandId}/retry:
    post:
      operationId: retrySensorCommand
      summary: Replace a dead-lettered command with a new signed delivery
      security:
        - controlPlaneSession: []
        - controlPlaneBearer: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/CommandId'
      responses:
        '201': { description: Replacement command queued }
        '404': { description: Command not found }
        '409': { description: Command is not dead-lettered }
  /sensor-commands/{commandId}/dismiss:
    post:
      operationId: dismissSensorCommand
      summary: Dismiss a dead-lettered command after operator review
      security:
        - controlPlaneSession: []
        - controlPlaneBearer: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/CommandId'
      responses:
        '200': { description: Command dismissed }
        '404': { description: Command not found }
        '409': { description: Command is not dead-lettered }
  /sensor-enrollment-tokens:
    post:
      operationId: createSensorEnrollmentToken
      summary: Create a single-use projection-sensor enrollment token
      security:
        - controlPlaneSession: []
        - controlPlaneBearer: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, maxLength: 100 }
                ttlMinutes: { type: integer, minimum: 5, maximum: 60, default: 15 }
      responses:
        '201': { description: Single-use token created }
        '401': { description: Control-plane authentication required }
  /sensors/enroll:
    post:
      security: []
      operationId: enrollSensor
      summary: Exchange a single-use token for a projection-sensor identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollmentToken, name]
              properties:
                enrollmentToken: { type: string }
                sensorId: { type: string, minLength: 3, maxLength: 64 }
                name: { type: string, maxLength: 100 }
                version: { type: string }
                platform: { type: string }
                capabilities:
                  type: array
                  items: { type: string }
      responses:
        '201': { description: Sensor enrolled; credentials are returned once }
        '401': { description: "Token invalid, expired, or used" }
        '409': { description: Sensor ID already exists }
  /sensors/{sensorId}/heartbeat:
    post:
      operationId: reportSensorHeartbeat
      summary: Report sensor health and active decoy inventory
      security:
        - sensorBearer: []
      parameters:
        - $ref: '#/components/parameters/SensorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                health: { type: number, minimum: 0, maximum: 100 }
                latency: { type: number, minimum: 0 }
                version: { type: string }
                address: { type: string }
                decoys:
                  type: array
                  items: { type: object }
      responses:
        '200': { description: Heartbeat accepted }
        '401': { description: Sensor authentication required }
  /sensors/{sensorId}/commands:
    get:
      operationId: pollSensorCommands
      summary: Poll queued, signed, unexpired commands
      security:
        - sensorBearer: []
      parameters:
        - $ref: '#/components/parameters/SensorId'
      responses:
        '200': { description: Signed command collection }
        '401': { description: Sensor authentication required }
    post:
      operationId: createSensorCommand
      summary: Queue a signed command for a projection sensor
      security:
        - controlPlaneSession: []
        - controlPlaneBearer: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/SensorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [deploy_decoy, stop_decoy, snapshot] }
                payload: { type: object }
                maxAttempts: { type: integer, minimum: 1, maximum: 10, default: 3 }
                retryBackoffSeconds: { type: integer, minimum: 5, maximum: 300, default: 15 }
      responses:
        '201': { description: Signed command queued }
        '404': { description: Sensor not found }
  /sensors/{sensorId}/commands/{commandId}/ack:
    post:
      operationId: acknowledgeSensorCommand
      summary: Acknowledge successful or failed sensor command execution
      security:
        - sensorBearer: []
      parameters:
        - $ref: '#/components/parameters/SensorId'
        - name: commandId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [acknowledged, failed] }
                output: {}
                error: { type: string }
      responses:
        '200': { description: Command state updated }
        '404': { description: Command not found }
  /sensors/{sensorId}/events:
    post:
      operationId: ingestSensorEvent
      summary: Convert a projected-decoy interaction into an incident
      security:
        - sensorBearer: []
      parameters:
        - $ref: '#/components/parameters/SensorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decoyId, source, protocol]
              properties:
                decoyId: { type: string }
                decoyName: { type: string }
                protocol: { type: string, enum: [http, ssh, postgres, redis, smb, tcp] }
                source: { type: string }
                destination: { type: string }
                metadata: { type: object }
      responses:
        '201': { description: Event and high-confidence incident created }
        '401': { description: Sensor authentication required }
...
components:
...
  securitySchemes:
    controlPlaneSession:
      type: apiKey
      in: cookie
      name: __Host-aidecepticon.sid
      description: HTTP-only server-side operator session established through OIDC, SAML, or API-key exchange.
    controlPlaneBearer:
      type: http
      scheme: bearer
      description: Optional control-plane API key configured by the administrator.
    sensorBearer:
      type: http
      scheme: bearer
      description: Unique sensor access token returned once during enrollment.
  parameters:
    OrganizationId:
      name: X-AIDecepticon-Organization
      in: header
      required: false
      description: Authorized organization to make active. The server selects the configured default when omitted.
      schema: { type: string, pattern: '^org-' }
    SensorId:
      name: sensorId
      in: path
      required: true
      schema: { type: string }
    CommandId:
      name: commandId
      in: path
      required: true
      schema: { type: string }
...